fixed Preis return

+++ domainLogic/src/kuchenImp/KremkuchenImp.java
@@ -26,7 +26,7 @@ public class KremkuchenImp extends KuchenImp implements Kremkuchen, Serializable
         return "Fachnummer: " + this.Fachnummer +
                 " | KuchenTyp: " + this.KuchenTyp +
                 " | Hersteller: " + this.Hersteller.getName() +
-                " | Preis: " + this.Preis +
+                " | Preis: " + this.Preis.doubleValue() +
                 " | Nährwert: " + this.Naehrwert +
                 " | verbleibende Haltbarkeit: " + this.getRemainingHaltbarkeit().toDays() + ((this.getRemainingHaltbarkeit().toDays() == 1L) ? " Tag" : " Tage") +
                 " | Allergen: " + this.Allergene +
