+++ domainLogic/src/kuchenImp/ObstkuchenImp.java
@@ -27,7 +27,7 @@ public class ObstkuchenImp extends KuchenImp implements Obstkuchen, Serializable
         return "Fachnummer: " + this.Fachnummer +
                 " | KuchenTyp: " + this.KuchenTyp +
                 " | Hersteller: " + this.Hersteller.getName() +
-                " | Preis: " + this.Preis +
+                " | Preis: " + this.Preis.doubleValue() +
                 " | Nährwert: " + this.Naehrwert +
                 " | verbleibende Haltbarkeit: " + this.getRemainingHaltbarkeit().toDays() + ((this.getRemainingHaltbarkeit().toDays() == 1L) ? " Tag" : " Tage") +
                 " | Allergen: " + this.Allergene +
